refactor(api): migrate sco import to src bounded context
