+++ api/lib/application/organizations/index.js
@@ -666,45 +666,6 @@ const register = async function (server) {
         tags: ['api', 'organization-learners'],
       },
     },
-    {
-      method: 'POST',
-      path: '/api/organizations/{id}/sco-organization-learners/import-siecle',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsAdminInSCOOrganizationManagingStudents,
-            assign: 'isAdminInOrganizationManagingStudents',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-          query: Joi.object({
-            format: Joi.string().default('xml'),
-          }),
-        },
-        payload: {
-          maxBytes: TWENTY_MEGABYTES,
-          output: 'file',
-          failAction: (request, h) => {
-            return sendJsonApiError(
-              new PayloadTooLargeError('An error occurred, payload is too large', ERRORS.PAYLOAD_TOO_LARGE, {
-                maxSize: '20',
-              }),
-              h,
-            );
-          },
-        },
-        handler: organizationController.importOrganizationLearnersFromSIECLE,
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés et responsables de l'organisation**\n" +
-            "- Elle permet d'importer des inscriptions d'élèves, en masse, depuis un fichier au format XML ou CSV de SIECLE\n" +
-            '- Elle ne retourne aucune valeur de retour',
-        ],
-        tags: ['api', 'organization-learners'],
-      },
-    },
   ];
 
   server.route([...adminRoutes, ...orgaRoutes]);
