+++ api/lib/domain/usecases/index.js
@@ -110,8 +110,6 @@ import * as organizationInvitedUserRepository from '../../infrastructure/reposit
 import * as organizationLearnerActivityRepository from '../../infrastructure/repositories/organization-learner-activity-repository.js';
 import * as organizationLearnerFollowUpRepository from '../../infrastructure/repositories/organization-learner-follow-up/organization-learner-repository.js';
 import * as organizationLearnerRepository from '../../infrastructure/repositories/organization-learner-repository.js';
-import * as organizationLearnersCsvService from '../../domain/services/organization-learners-csv-service.js';
-import * as organizationLearnersXmlService from '../../domain/services/organization-learners-xml-service.js';
 import * as organizationMemberIdentityRepository from '../../infrastructure/repositories/organization-member-identity-repository.js';
 import * as organizationRepository from '../../infrastructure/repositories/organization-repository.js';
 import * as organizationsToAttachToTargetProfileRepository from '../../infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
@@ -313,8 +311,6 @@ const dependencies = {
   organizationLearnerActivityRepository,
   organizationLearnerFollowUpRepository,
   organizationLearnerRepository,
-  organizationLearnersCsvService,
-  organizationLearnersXmlService,
   organizationMemberIdentityRepository,
   organizationRepository,
   organizationsToAttachToTargetProfileRepository,
