+++ api/src/shared/infrastructure/constants.js
@@ -2,8 +2,11 @@ import { config } from '../config.js';
 
 const CONCURRENCY_HEAVY_OPERATIONS = config.infra.concurrencyForHeavyOperations;
 
+const ORGANIZATION_LEARNER_CHUNK_SIZE = config.infra.chunkSizeForOrganizationLearnerDataProcessing;
+
 const constants = {
   CONCURRENCY_HEAVY_OPERATIONS,
+  ORGANIZATION_LEARNER_CHUNK_SIZE,
 };
 
-export { constants, CONCURRENCY_HEAVY_OPERATIONS };
+export { constants, CONCURRENCY_HEAVY_OPERATIONS, ORGANIZATION_LEARNER_CHUNK_SIZE };
