+++ api/tests/unit/application/organizations/index_test.js
@@ -737,47 +737,6 @@ describe('Unit | Router | organization-router', function () {
     });
   });
 
-  describe('POST /api/organizations/{id}/sco-organization-learners/import-siecle', function () {
-    it('should call the organization controller to import organizationLearners', async function () {
-      // given
-      const method = 'POST';
-      const url = '/api/organizations/1/sco-organization-learners/import-siecle';
-      const payload = {};
-
-      sinon
-        .stub(securityPreHandlers, 'checkUserIsAdminInSCOOrganizationManagingStudents')
-        .callsFake((request, h) => h.response(true));
-      sinon
-        .stub(organizationController, 'importOrganizationLearnersFromSIECLE')
-        .callsFake((request, h) => h.response('ok').code(201));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(201);
-      expect(organizationController.importOrganizationLearnersFromSIECLE).to.have.been.calledOnce;
-    });
-
-    it('should throw an error when id is invalid', async function () {
-      // given
-      const method = 'POST';
-      const url = '/api/organizations/wrongId/sco-organization-learners/import-siecle';
-      const payload = {};
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-
   describe('DELETE /api/organizations/{id}/invitations/{invitationId}', function () {
     it('should call the cancel organization invitation controller', async function () {
       // given
